Fix: when inventory is full, cannot move from bottom to top
That is: point to the 127th element, try to go down (to wraparound and
go on top); nothing will happen.

+++ src/scene/chest.c
@@ -22,7 +22,7 @@
 #include "player.h"
 #include "furniture.h"
 
-static i8 chest_selected[2] = { 0, 0 };
+static i32 chest_selected[2] = { 0, 0 };
 static u8 chest_window;
 
 static struct Inventory *chest_inventory;
